requirements.txt
requests
pandas
plotly
ccxt
google-generativeai
duckduckgo-search
beautifulsoup4
pypdf
streamlit
uvicorn
fastapi
jupyter
sqlalchemy
psycopg2-binary
apscheduler
yfinance
webdriver_manager
holidays
pydantic
selenium
python-dateutil